+++ src/ui/login_view.py
@@ -1,9 +1,18 @@
 from tkinter import ttk, constants
-from services.user_service import user_service, InvalidCredentialsError
+from services.user_service import user_service
 
 
 class LoginView:
+    """View responsible for user logins
+    """
     def __init__(self, root, handle_login, handle_show_create_user_view):
+        """Constructor for the class. Creates new LoginView
+
+        Args:
+            root : tkInter element where the ui will be outlined
+            handle_login: Value that is called when user logs in
+            handle_show_create_user_view: Value that is called when user wants to see create user view
+        """
         self._root = root
         self._username_entry = None
         self._password_entry = None
@@ -14,9 +23,13 @@ class LoginView:
         self._initialize()
 
     def pack(self):
+        """Shows the view
+        """
         self._frame.pack(fill=constants.X)
 
     def destroy(self):
+        """Destroys the view
+        """
         self._frame.destroy()
 
     def _initialize_username_field(self):
@@ -32,8 +45,31 @@ class LoginView:
         password_label.grid(sticky=constants.W, padx=5, pady=5)
         self._password_entry.grid(
             row=2, column=1, sticky=constants.EW, padx=5, pady=5)
+    
+    def _initialize_login_button(self):
+        login_button = ttk.Button(
+            master=self._frame,
+            text="Login",
+            command=self._login_handler
+        )
+
+        login_button.grid(sticky=constants.W, padx=5, pady=5)
+
+    def _initialize_create_user_button(self):
+        create_user_button = ttk.Button(
+            master=self._frame,
+            text="Create new user",
+            command=self._handle_show_create_user_view,
+        )
+
+        create_user_button.grid(
+            row=3, column=1, sticky=constants.W, padx=5, pady=5)
+
 
     def _initialize(self):
+        """Creates the frame where the login view is build with the help of other functions
+        """
+
         self._frame = ttk.Frame(master=self._frame)
 
         header = ttk.Label(master=self._frame, font=24, text="Login")
@@ -41,21 +77,8 @@ class LoginView:
 
         self._initialize_username_field()
         self._initialize_password_field()
-
-        login_button = ttk.Button(
-            master=self._frame,
-            text="Login",
-            command=self._login_handler
-        )
-        create_user_button = ttk.Button(
-            master=self._frame,
-            text="Create new user",
-            command=self._handle_show_create_user_view,
-        )
-
-        login_button.grid(sticky=constants.W, padx=5, pady=5)
-        create_user_button.grid(
-            row=3, column=1, sticky=constants.W, padx=5, pady=5)
+        self._initialize_login_button()
+        self._initialize_create_user_button()
         self._frame.grid_columnconfigure(1, weight=1, minsize=300)
 
     def _login_handler(self):
@@ -65,5 +88,5 @@ class LoginView:
         username = self._username_entry.get()
         password = self._password_entry.get()
 
-        user_service.login(username, password)
-        self._handle_login()
+        if user_service.login(username, password):
+            self._handle_login()